+++ packages/plugin-text-color/web/src/textHighlightInlineStyleMapper.jsx
@@ -1,26 +1,26 @@
 import React from 'react';
-import { DEFAULT_STYLE_SELECTION_PREDICATE, DEFAULT_FOREGROUND_STYLE_FN } from './constants';
-import { TEXT_COLOR_TYPE } from './types';
+import { DEFAULT_STYLE_SELECTION_PREDICATE, DEFAULT_BACKGROUND_STYLE_FN } from './constants';
+import { TEXT_HIGHLIGHT_TYPE } from './types';
 import {
-  textForegroundPredicate,
+  textBackgroundPredicate,
   viewerStyleFnFilter,
-  isTextColor,
+  isTextHighlight,
 } from './text-decorations-utils';
 
 /**
- * textColorInlineStyleMapper
+ * textHighlightInlineStyleMapper
  * @param {object} config consumer plugin configuration
  * @param {object} raw raw content state
  * @returns {function} mapping of inline style => component
  */
 export default (config, raw = { blocks: [] }) => {
-  const settings = config[TEXT_COLOR_TYPE] || {};
-  const styleSelectionPredicate = textForegroundPredicate(
+  const settings = config[TEXT_HIGHLIGHT_TYPE] || {};
+  const styleSelectionPredicate = textBackgroundPredicate(
     settings.styleSelectionPredicate || DEFAULT_STYLE_SELECTION_PREDICATE
   );
   const customStyleFn =
-    (settings.customStyleFn && viewerStyleFnFilter(settings.customStyleFn, isTextColor)) ||
-    DEFAULT_FOREGROUND_STYLE_FN;
+    (settings.customStyleFn && viewerStyleFnFilter(settings.customStyleFn, isTextHighlight)) ||
+    DEFAULT_BACKGROUND_STYLE_FN;
   const mapper = raw.blocks.reduce((map, block) => {
     if (block.inlineStyleRanges) {
       block.inlineStyleRanges