+++ packages/plugin-text-color/web/src/toolbar/TextHighlightButton.jsx
@@ -1,32 +1,31 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import TextColorIcon from './TextColorIcon';
-import { TEXT_COLOR_TYPE } from '../types';
+import TextHighlightIcon from './TextHighlightIcon';
+import { TEXT_HIGHLIGHT_TYPE } from '../types';
 import BaseTextColor from './BaseTextColor';
-import { textForegroundPredicate } from '../text-decorations-utils';
+import { textBackgroundPredicate } from '../text-decorations-utils';
 import { get } from 'lodash';
 
-export default class TextColorButton extends Component {
+export default class TextHighlightButton extends Component {
   constructor(props) {
     super(props);
     this.buttonRef = React.createRef();
   }
-
   render() {
-    const settings = this.props.config[TEXT_COLOR_TYPE];
-    const icon = get(settings, 'toolbar.icons.TextColor');
+    const settings = this.props.config[TEXT_HIGHLIGHT_TYPE];
+    const icon = get(settings, 'toolbar.icons.TextHighlight');
     const pluginParams = {
-      dataHook: 'TextColorButton',
-      toolTip: 'TextColorButton_Tooltip',
-      icon: icon || TextColorIcon,
-      type: TEXT_COLOR_TYPE,
-      predicate: textForegroundPredicate,
+      dataHook: 'TextHighlightButton',
+      toolTip: 'TextHighlightButton_Tooltip',
+      icon: icon || TextHighlightIcon,
+      type: TEXT_HIGHLIGHT_TYPE,
+      predicate: textBackgroundPredicate,
     };
     return <BaseTextColor buttonRef={this.buttonRef} pluginParams={pluginParams} {...this.props} />;
   }
 }
 
-TextColorButton.propTypes = {
+TextHighlightButton.propTypes = {
   getEditorState: PropTypes.func.isRequired,
   setEditorState: PropTypes.func.isRequired,
   onExtendContent: PropTypes.func.isRequired,
@@ -45,6 +44,6 @@ TextColorButton.propTypes = {
   setKeepOpen: PropTypes.func,
 };
 
-TextColorButton.defaultProps = {
+TextHighlightButton.defaultProps = {
   setKeepOpen: () => {},
 };