+++ packages/plugin-text-color/web/src/viewer.js
@@ -1,2 +1,3 @@
-export { TEXT_COLOR_TYPE } from './types';
+export { TEXT_COLOR_TYPE, TEXT_HIGHLIGHT_TYPE } from './types';
 export { default as textColorInlineStyleMapper } from './textColorInlineStyleMapper';
+export { default as textHighlightInlineStyleMapper } from './textHighlightInlineStyleMapper';